*db/Object.php: Added permissions system *db/Object.php (isAllowed): New function to check permissions. *db/Object.php (getInfo): New function for getting properties. *db/Object.php (putInfo): New function for setting properties.
Added a system to give permissions for accessing one object from another.
Groups for an object are stored in the auth column of the database table, and
permissions are derived from the groups using the configuration file.

When accessing another object, the DB_Object being accessed calls isAllowed
on the other object to see if it is allowed to do something.

Signed-off-by: Tyler Romeo <parent5446@parent5446-desktop.(none)>